chore: cycles quick actions restructuring
