+++ web/ce/components/cycles/index.ts
@@ -1,2 +1,3 @@
 export * from "./active-cycle";
 export * from "./analytics-sidebar";
+export * from "./quick-actions";
