+++ web/ee/components/cycles/index.ts
@@ -1,2 +1,3 @@
 export * from "./active-cycle";
 export * from "./analytics-sidebar";
+export * from "ce/components/cycles/quick-actions";
